+++ data/src/main/kotlin/us/wedemy/eggeum/android/data/model/File.kt
@@ -10,19 +10,10 @@ package us.wedemy.eggeum.android.data.model
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
-// @JsonClass(generateAdapter = true)
-// public data class File(
-//   @Json(name = "uploadFileId")
-//   val uploadFileId: Int,
-//
-//   @Json(name = "url")
-//   val url: String,
-// )
-
 @Serializable
 public data class File(
   @SerialName("uploadFileId")
-  val uploadFileId: Int?,
+  val uploadFileId: Long?,
 
   @SerialName("url")
   val url: String?,
