+++ domain/src/main/kotlin/us/wedemy/eggeum/android/domain/model/notice/NoticeEntity.kt
@@ -8,7 +8,7 @@
 package us.wedemy.eggeum.android.domain.model.notice
 
 public data class NoticeEntity(
-  val id: Int,
+  val id: Long,
   val title: String,
   val content: String,
   val viewCount: Int,
